Handle page breaks in BRF (#163)
`splitlines()` was eating `\f` so we never got to act on them.
Since `re.match()` is implicitly anchored to the start of the line,
also remove the case of embedded `\f` which never fires.

This leaves in the air the question of how to handle e.g. "foo\fbar",
but since BRF doesn't appear to have a formal spec anywhere and since
we only ever see `\f` immediately following end-of-line in any of our
example BRFs, put off handling that unless/until we see it for real.