Bugfix: catch all mounts & handle SD differences
Subscribing to a block device, catching a change, then responding
by subscribing to mounts of that device was too slow and we'd often
miss the mount event.

Instead subscribe to all mount events and filter for interesting
ones.

Use the same style to watch for the different manifestation
of SD card media presence changes (compared to USB stick changes).