Serial: enforce lockstep using CMD_POLL
Historically SEND_LINE was synchronous and we'd get lockstep by waiting
for the response before sending the next one.  Then firmware changed to
reply immediately and perform motion in the background, while we didn't
change; in practice the firmware would buffer the following request
while it actuated, so it didn't matter.  As of now, to deliver a sane
protocol for Canute-as-device scenarios, the firmware is changing to be
consistent across the commands and across repeated issue of commands,
so the UI must either enforce lockstep or handle temporary failures.