Retry unacknowledged SEND_LINEs
Without some sort of timeout we'd hang indefinitely.  Embarrassing if
a user suffers that, but potentially terrible if it happens during EMC
bombardment and we mistakenly conclude that bombardment caused a hang.

This should hopefully go away if we ever improve the firmware and
protocol to have instant responses under all conditions.